fix(dns): segfaults in pkarr (#3120)
## Description

`simple-dns` had a number of risky parsing calls that could go out of
bounds and produce a segfault. Through the power of OSS and the daisy
chain of bumped dependencies this should now be fixed.

Closes https://github.com/n0-computer/iroh/issues/2844

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.