chore: Remove checkboxes from PR templates
A significant number of our own PRs don't bother with the checkboxes.
This makes it hard to enforce (I've tried!) and does not give us
anything to ask from external contributors either.

I'm also often torn what to do with non-applicable items, I've tried
all of:
- Remove them.  Weird.
- Strike them out, not so weird, but so much effort.
- Tick off even if not-applicable (my current approache), also weird.

So instead of bothering the humans I'm adapting to their behaviour.

An alternative I could live with would be enforce this by machines:
Have a required CI check that needs the checkboxes ticked off so that
you can not merge a PR while ignoring this.  Maybe with some more
guidance on what to do if not applicable.