fix(iroh): Return error if disco send via relay fails
The signature of this function expects an error if the send fails, but
the relay part did not do this.  The impact of this is that you might
silently drop a packet, which makes debugging a bit harder.  Returning
an error should propagate this correctly and the failures will be
noticed and logged.