refactor(iroh): Replace `timer` module with `AbortOnDropHandle` and sleep (#3141)
## Description

I find this easier to reason about. Better this than having to
understand an additional concept (the golang timer).

## Notes & open questions

This is opinionated. I'm not married to this, so feel free to critique.

This also makes it such that when the MagicSocket is dropped (/the
NodeMap is dropped), it'll abort all these tasks.

## Change checklist

- [x] Self-review.