chore: new project syncing setup (#3136)
## Description
Revamps how we track PRs and Issues to add onto repos. Mainly because
this produced issues with forks but also did all the work attribution to
me.

The downside of this approach is that it runs on a schedule (set to
every 6 hours but we can change). Also if GH api craps out on us we have
to wait till the next trigger.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.