feat(iroh-relay): Make the client side of `iroh-relay` compile & run in browsers (#3119)
## Description

<!-- A summary of what this pull request achieves and a rough list of
changes. -->
Allows compiling and running `iroh-relay` in browsers (as long as the
`server` feature is disabled).

By default uses websockets to connect to relays when compiling for the
`wasm32-unknown-unknown` target.
If configured to `Protocol::Relay` specifically, it'll error out.

Test it yourself: `cargo rustc -p iroh-relay --no-default-features
--target=wasm32-unknown-unknown --crate-type=cdylib`

## Breaking Changes

- iroh: `iroh-quinn` re-exports are now updated from version 0.12 to
version 0.13

## Notes & open questions

**Note:** This updates *all* crate's versions to `iroh-quinn` 0.13!
(Because we need them to be consistent between crates in this
workspace.)

## TODO

- [x] Add CI check to make sure the built .wasm is `(import "env"
...)`-free
- [x] Need a new release of `iroh-quinn` with my browser PR in
- [x] ~~**Blocking this from merging**: Need a new `netwatch` release
with the `quinn-udp = "0.5.5"` version bound relaxed~~ it's already
relaxed

## Change checklist

- [x] Self-review.
- [x] All breaking changes documented.