feat(iroh-net-report): Support wasm32 building & running (#3139)
## Description

This makes iroh-net-report compile to Wasm. We make sure it's likely to
work by grepping the generated Wasm file for "env" imports (which would
make it not work in browsers).
Mostly this will disable functionality in iroh-net-report for the
browser target - specifically, it disables STUN, QAD, ICMP and all other
probes besides HTTPS to relays.
I've experimentally verified this code works with an integration test,
it's really hard to come up with a good strategy for automating this
testing due to requiring running cross-languages.
I'd like to do that as a follow-up once #3145 is merged.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.