fix(iroh): Allow gracefully closing connections (#3170)
## Description

This adds a test case that checks that we can call `Connection::close`
and actually receive the close code on the other side, indicating we've
gracefully closed the connection.
Even though in the test we call `Endpoint::close` and wait for that,
with current iroh 0.32, we don't receive the close frame. The only way I
can see to make it work with iroh's 0.32 API, is to choose some time to
sleep before calling `Endpoint::close`.

I've also attached a revert of #3165 which fixes the test, as that makes
`Endpoint::close` wait for the close frames to be acknowledged again (by
calling `quinn::Endpoint::wait_idle`).

## Notes & open questions

There might be other ways to fix this, I welcome feedback on that. I do
believe this is the right way to fix it, though.

Given that the `Endpoint::wait_idle` call has been introduced and
removed *twice* already, we should finally make a choice on whether it
belongs in there or not, and then add sufficient code comments to make
sure whoever looks at it next is fully informed.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.