feat(iroh): add subscription stream to watch all discovery results (#3181)
## Description

Implements a subscription stream to watch all items retrieved through
discovery services. the stream includes both nodes discovered actively
(through `Discovery::resolve`, which is invoked when `Endpoint::connect`
is called for nodes we don't have addressing info), and nodes discovered
passively (through `Discovery::subscribe`, which is only implemented for
the local swarm discovery).

Uses `tokio::sync:.broadcast` internally (but without exposing this in
the public API) which fits well here I think: if a consumer is too slow
in processing the stream, it will miss events but can still continue.

Currently the capacity of the broadcast channel is hardcoded to 128. We
could also expose this as a setting on the endpoint builder, not sure if
there is a default that fits everything.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.