fix(iroh): Reap ActiveRelayActor handles for idle relays (#3249)
## Description

The ActiveRelayActor exists when it is idle for a while (60s for now).
When it does the RelayActor is supposed to notice the tasks exists and
clean up the handles it keeps to all the ActiveRelayActors.
Unfortunately it was only doing this for ActiveRelayActors which
errored and not for those that exited normally because they were idle.

This makes sure to also reap active relay handles when the actor
exists normally.

Additionally this stops propagating the panic.  The relay actor itself
is currently spawned without keeping track of the handle.  Panicking
it would make it disappear silently.  Propagating panics is
reasonable, but I don't want to fix the task spawning.  That's for
another time.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

I looked at writing a test for this, but testing the RelayActor is
still something that's difficult to do.  So sadly no tests.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.