feat(iroh): move `iroh-net-report` back into `iroh` (#3251)
## Description

Move `iroh-net-report` back into `iroh`

Only the follow are made public from the old `iroh-net-report` (now
`iroh::net_report`):
- `iroh::net_report::Metrics`
- `iroh::net_report::Options`
- `iroh::net_report::Addr`
- `iroh::net_report::Client`
- `iroh::net_report::Report`
- `iroh::net_report::bind_local_stun_addr`
- `iroh::net_report::RelayLatencies`
- `iroh::net_report::QuicConfig`

## Breaking Changes

Technically not a breaking change to `iroh` since we are "adding" to the
API, but we will no longer be publishing `iroh-net-report` to crates.io,
and to use the content from `iroh-net-report` you would need to depend
on `iroh` and import it using `iroh::net_report`.

This should only effect `iroh-doctor`

## Notes & open questions

Would be nice to make this private, if we can adjust `iroh-doctor`
accordingly.

For the next release, I'll make a branch that has the old
`iroh-net-report`, adds the deprecation status to the readme/cargo.toml
& publishes.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.

---------

Co-authored-by: “ramfox” <“kasey@n0.computer”>