fix: backoff before retry if relay connection terminates (#3254)
## Description

Currently, if the relay server aborts a connection, we would reconnect
to the relay server right away. If the relay server happens to always
abort the connection right after establishing, we would be retrying in a
hot loop without pause. This happens reliably if the relay server aborts
the connection because it rejects the client as not authorized.

This PR has two changes:

* The existing backoff for retrying failed dials is move up a layer, so
that we also backoff after the connection closes before retrying.
* If the server reports a `Health` problem, which it does for
unauthenticated clients right before closing the connection, this is
logged with `WARN` level

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [ ] All breaking changes documented.
- [ ] List all breaking changes in the above "Breaking Changes" section.
- [ ] Open an issue or PR on any number0 repos that are affected by this
breaking change. Give guidance on how the updates should be handled or
do the actual updates themselves. The major ones are:
    - [ ] [`quic-rpc`](https://github.com/n0-computer/quic-rpc)
    - [ ] [`iroh-gossip`](https://github.com/n0-computer/iroh-gossip)
    - [ ] [`iroh-blobs`](https://github.com/n0-computer/iroh-blobs)
    - [ ] [`dumbpipe`](https://github.com/n0-computer/dumbpipe)
    - [ ] [`sendme`](https://github.com/n0-computer/sendme)