feat(iroh-relay): allow to authenticate nodes via a HTTP POST request (#3246)
## Description

This adds a new `access` config option to the `iroh-relay` binary: If
set to `http`, a `POST` request will be performed for each node that
attempts to connect to the relay. The URL to request is set in the
config, and a `X-Iroh-NodeId` header will be set to the hex-encoded node
id to check access for. If and only if the response has a 200 status
code, and a response text of `true`, the connecting node will be granted
access. In all other cases, the node will be denied access.

Config example to use this:
```toml
access.http.url = "http://localhost:8000/api/relays/check-auth/frando"
```

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.