chore(iroh): Add `echo-no-router.rs` example (#3267)
## Description

Adds an example showing how to implement a basic protocol without using
a `Router`.
As a bonus, one can compare what `echo.rs` and `echo-no-router.rs` are
doing to see the difference.

## Notes & open questions

I think this is slightly useful, but YMMV. I don't think it's a lot of
work to maintain though.

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.

---------

Co-authored-by: Floris Bruynooghe <flub@n0.computer>