fix(iroh-relay): don't stop relay client actor if queues become full (#3294)
## Description

This is a fix to the iroh-relay server.

Current situation: If a client stream becomes full (because the client
is not reading fast enough from the network), the internal queue where
messages to that client are queued becomes full too. Now, when another
client wants to send a message to the filled-up client, instead of
skipping over the message we propagate the error such that the sending
client's actor dies. This is bad, we should instead just drop the
message.

This PR changes the behavior to log the failed forward (at WARN level)
but does not stop the sending client's actor.

It also adds a test that fails without the change and passes with the
change.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Tests if relevant.