fix(iroh-dns-server): Backwards compatibility for packets stored with iroh-dns-server v0.34 or lower (#3295)
## Description

#3186 introduced a backwards-incompatible change to the iroh-dns-server
database that went uncaught so far.
The serialization format for SignedPackets changed between pkarr v2 and
pkarr v3: pkarr v3 prepends a `last_seen` timestamp (8 bytes) to the
serialized representation of a SignedPacket. Therefore, parsing packets
stored in the the database with pkarr v2 (iroh-dns-server v0.34 and
lower) fails with an error. The error also stops the store actor, and
thus all subsequent requests will fail as well.

This PR changes this by attempting to parse the packet in the pkarr v2
format if parsing as pkarr v3 packet fails. Parsing as pkarr v2 is
achieved by prepending 8 zero bytes to the payload (an empty last_seen
timestamp).

The PR also improves the tracing logging at various places.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

There's no test yet. I tested manually by launching iroh-dns-server
v0.34, inserting some packets with the `publish` example. When
restarting the server from `main`, retrieving packets leads to errors.
When applying this PR, retrieving the old packets work.

If we want / need a test, how should we do it? We can either add a dev
dependency to `iroh_dns_server@0.34` (which would recursively pull in
all of `iroh@0.34`), or somehow insert a packet as it would be created
by `iroh_dns_server@0.34` manually into the database. Not sure if it's
worth it, but maybe it is?

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.