chore(iroh-relay): Make QAD test non-flaky by using tokio's paused time (#3341)
## Description

Use `tokio`'s paused time to make the QAD test non-flaky. We've
experienced occasional slowdowns recently every now and then, causing
our timeout test to get out of the "acceptable" range.
Instead, this test now pauses the time, and will assert that the time
that we sleep before considering the datagrams lost forever to be
*exactly* 999ms.

## Notes & open questions

Doesn't "real world" test the code, but honestly, this is testing that a
timeout is set fairly low, so I think this should be acceptable?

## Change checklist
<!-- Remove any that are not relevant. -->
- [x] Self-review.